# shared/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Smart Parking Management System API",
    "version": "1.0.0",
    "description": "REST API contract for Smart Parking Management System per Frozen Contract V1"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4000"
    }
  ],
...
  "paths": {
...
    "/api/v1/admin/lots": {
      "post": {
        "summary": "Create parking lot",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "address": {
                    "type": "string",
                    "minLength": 1
                  },
                  "latitude": {
                    "type": "number",
                    "minimum": -90,
                    "maximum": 90
                  },
                  "longitude": {
                    "type": "number",
                    "minimum": -180,
                    "maximum": 180
                  },
                  "totalSlots": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "pricePerHourPaise": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  }
                },
                "required": [
                  "name",
                  "address",
                  "latitude",
                  "longitude",
                  "totalSlots",
                  "pricePerHourPaise"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Parking lot created"
          }
        }
      }
    },
    "/api/v1/admin/lots/{id}": {
      "put": {
        "summary": "Update parking lot",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "address": {
                    "type": "string",
                    "minLength": 1
                  },
                  "latitude": {
                    "type": "number",
                    "minimum": -90,
                    "maximum": 90
                  },
                  "longitude": {
                    "type": "number",
                    "minimum": -180,
                    "maximum": 180
                  },
                  "totalSlots": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "pricePerHourPaise": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Parking lot updated"
          }
        }
      },
      "delete": {
        "summary": "Deactivate parking lot",
        "responses": {
          "200": {
            "description": "Parking lot deactivated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "deactivated": {
                          "type": "boolean"
                        },
                        "lotId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "deactivated",
                        "lotId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/guard/check-in": {
      "post": {
        "summary": "Check in a booking by booking code",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bookingCode": {
                    "type": "string",
                    "pattern": "^[ABCDEFGHJKLMNPQRSTUVWXYZ23456789]{6}$"
                  }
                },
                "required": [
                  "bookingCode"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Booking checked in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "slotId": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "HELD",
                            "CONFIRMED",
                            "CANCELLED",
                            "EXPIRED",
                            "COMPLETED",
                            "NO_SHOW"
                          ]
                        },
                        "startTime": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "endTime": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "amountPaise": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "heldUntil": {
                          "type": "string",
                          "nullable": true,
                          "format": "date-time"
                        },
                        "bookingCode": {
                          "type": "string"
                        },
                        "vehicleNumber": {
                          "type": "string",
                          "nullable": true
                        },
                        "checkedInAt": {
                          "type": "string",
                          "nullable": true,
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "id",
                        "slotId",
                        "status",
                        "startTime",
                        "endTime",
                        "amountPaise",
                        "heldUntil",
                        "bookingCode",
                        "vehicleNumber",
                        "checkedInAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/guard/slots/{slotId}/walk-in": {
      "post": {
        "summary": "Toggle walk-in status on slot",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "OCCUPIED",
                      "AVAILABLE"
                    ]
                  }
                },
                "required": [
                  "status"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Walk-in status updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "slotId": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "AVAILABLE",
                            "HELD",
                            "RESERVED",
                            "OCCUPIED"
                          ]
                        },
                        "source": {
                          "type": "string",
                          "enum": [
                            "GUARD"
                          ]
                        }
                      },
                      "required": [
                        "slotId",
                        "status",
                        "source"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  }
}